backend/requirements.txt
# Core Dependencies - as specified in PROJECT_REFERENCE.md
livekit-agents==0.8.3      # LiveKit agent framework (updated version)
livekit-api==0.6.0         # LiveKit API client for token generation (compatible with agents)
google-generativeai>=0.3.0 # Gemini AI response generation
mem0ai==0.0.8              # Memory and context storage
python-dotenv==1.0.0       # Environment management
requests==2.31.0           # HTTP client
aiohttp==3.9.0             # Async HTTP operations
flask==3.0.0               # Web server for token generation
flask-cors==4.0.0          # CORS support for frontend requests